AssertEqualsSpecializations trait: polyfill the Assert::assertEqualsWithDelta() et al methods
PHPUnit 7.5.0 introduced the new `Assert::assertEqualsCanonicalizing()`, `Assert::assertEqualsIgnoringCase()`, `Assert::assertEqualsWithDelta()`, `Assert::assertNotEqualsCanonicalizing()`, `Assert::assertNotEqualsIgnoringCase()` and `Assert::assertNotEqualsWithDelta()` methods, to replace use of the `Assert::assertEquals()` and `Assert::assertNotEquals()` methods with these optional parameters.
This type of use was soft deprecated in PHPUnit 7.5.0, hard deprecated (warning) in PHPUnit 8.0.0 and removed in PHPUnit 9.0.0.

This commit:
* Adds two traits with the same name.
    One to polyfill the methods when not available in PHPUnit.
    The other - an empty trait - to allow for `use`-ing the trait in PHPUnit versions in which the methods are already natively available.
* Logic to the custom autoloader which will load the correct trait depending on the PHPUnit version used.
* An availability test for the functionality polyfilled.

Note: the methods use `static::` to call the PHPUnit native functionality. This allows for existing method overloads in a child class of the PHPUnit native `TestCase` to be respected.