New `Helpers\AssertAttributeHelper`
The PHPUnit `assertAttribute*()` methods were deprecated in PHPUnit 8.x and removed in PHPUnit 9.0.

Public properties can still be tested by accessing them directly.
Protected and private properties can no longer be tested using PHPUnit native functionality.

The reasoning for the removal of these assertion methods is that _private and protected properties are an implementation detail and should not be tested directly, but via methods in the class_.

It is strongly recommended to refactor tests, and if needs be, classes to adhere to this.

However, if for some reason the value of protected or private properties still needs to be tested, this helper can be used to get access to their properties and value.

Includes tests.

Fixes 2