Composer: more specific PHPUnit requirements
This ensures that the "forward compatible" namespaced versions of common PHPUnit classes/interfaces are available.

PHP 6.0 introduced namespaces to PHPUnit.
The last few version(s) of PHPUnit 4.x and 5.x, contain both non-namespaced and namespaced versions of the most commonly referred to classes used by test suites.

By being strict about the minimum supported versions of PHPUnit 4.x and 5.x, we can be sure that the namespaced versions of the files are available when we need them and that we don't need to create `class_alias()`-es for these.