Allow installation in combination with PHPUnit 4.8.36
To allow for testing on PHP 5.5, PHPUnit 4 is needed.

PHPUnit 4.8.36 is the only version in the 4.x series which contains the forward-compatible layer with the namespaced versions of common PHPUnit classes.

With the additional polyfills which have been put in place, installation of this package on PHP 5.5 in combination with PHPUnit 4.8.36 is now supported.

Includes annotating this change in all the relevant places.
Includes adding a new build to the Travis matrix to also run all tests against PHP 5.5 / PHPUnit 4.8.36.