CI/QA: fix testVersion for PHPCompatibility
This library supports PHP >= 5.5, but the default `testVersion` as set in YoastCS is `5.6-`.
Overruling the `testVersion` from within a custom ruleset is currently not possible. This is a known issue in PHPCS itself and a fix is expected to be included with PHPCS 4.x.

In the mean time, as a work-around, the `testVersion` can be overruled from the command-line.