TestListenerTest: fix PHPUnit 10.0.0 compatibility
The `$name` parameter for the `TestCase::__construct()` method has become mandatory and should contain the name of the test method to run.

Ref: https://github.com/sebastianbergmann/phpunit/commit/705874f1b867fd99865e43cb5eaea4e6d141582f

---

This PHPUnit 10.0.0 change will generally speaking not affect "normal" users as it is rare for them to declare `__construct()` methods in test classes or to instantiate test classes directly, as this is the responsibility of PHPUnit.

In most cases, only test code testing for cross-version PHPUnit compatibility while using PHPUnit to run the tests will be affected.

Even so, fixing this is straight-forward, as the `$name` parameter was already an optional parameter since way back when, so passing the parameter in allows the tests to run (and pass) cross-version.