ExpectExceptionTest: fix PHPUnit 10.0.0 compatibility [1]
... by always passing a TestResult object to the `run()` method.

The `$testResult` parameter was previously optional, but will become required as of PHPUnit 10.0.0.
Also, the `run()` method used to return an instance of `TestResult`, but will now return `void`.

Ref: https://github.com/sebastianbergmann/phpunit/commit/77e60150c2185da760692277aeb2b337686b1f30

---

Along the same lines as 25, this PHPUnit 10.0.0 change will generally speaking not affect "normal" users.

In most cases, only test code testing for cross-version PHPUnit compatibility while using PHPUnit to run the tests will be affected.

Even so, fixing this is straight-forward, as the `$result` parameter was already an optional parameter since way back when, so instantiating the `TestResult` and passing the parameter in allows the tests to run cross-version.