AssertClosedResource trait: polyfill the Assert::assertIs[Not]ClosedResource() methods
PHPUnit 9.3.0 introduced the new `Assert::assertIsClosedResource()` and `Assert::assertIsNotClosedResource()` methods.

This commit:
* Adds two traits with the same name.
    One to polyfill the methods when not available in PHPUnit.
    The other to allow for `use`-ing the trait in PHPUnit versions in which the methods are already natively available.
* Logic to the custom autoloader which will load the correct trait depending on the PHPUnit version used.
* A `ResourceHelper` helper class containing the actual logic to determine whether something is a resource and whether that resource is open or closed.

As the PHP native functionality used in PHPUnit itself was only added to PHP in PHP 7.2.0 and the polyfills are available to older PHP versions, this polyfill needs actual logic and doesn't just fall through to an underlying PHPUnit native method.

With that in mind, an availability test is not sufficient to test this.
Additionally, resources in various extensions may behave and identify differently, so this commit also adds extensive tests with a variety of functionality using resources in PHP.

Note: the methods use `static::` to call the PHPUnit native functionality. This allows for existing method overloads in a child class of the PHPUnit native `TestCase` to be respected.

Includes:
* Adding information on the new polyfill to the README.
* Adding the new polyfill to the existing `TestCases` classes.
* Adding a few select exceptions  to the PHPCS ruleset.