AssertClosedResource: add `shouldClosedResourceAssertionBeSkipped()` helper function
While creating tests for this polyfill, I discovered that for resources created via the `libxml` extension, the PHP native functions showed some curious behaviour.

This behaviour basically makes the "is closed resource" determination unreliable for a select group of PHP versions in combination with resources created via the `libxml` extension.

To prevent tests failing on this, calls to the `assertIs[Not]ClosedResource()` method should - selectively - be wrapped in a condition checking the output of the `AssertClosedResource::shouldClosedResourceAssertionBeSkipped()` method, like so:
```php
if ( $this->shouldClosedResourceAssertionBeSkipped( $actual ) === false ) {
    $this->assertIsClosedResource( $actual );
}
```

As the PHPUnit native implementation of the assertions is also afflicted by the underlying bugs in PHP itself, this helper method is also available when the "empty" version of the `AssertClosedResource` trait is loaded and can therefore safely be used PHPUnit cross-version.

As things stand, the `libxml` inconsistency is the only one I have discovered so far.
However, there are quite a few extensions producing resources which are currently not covered in the tests as that would require databases or ssh/ftp/mail credentials to be added to the tests.

Users are encouraged to report other inconsistencies found and the `ResourceHelper::isResourceStateReliable()` method can be expanded with additional inconsistencies when found in the future.

Includes tests for this trait in combination with the `libxml` extension.
Also adds calls to a few of the other tests to ensure that the `shouldClosedResourceAssertionBeSkipped()` correctly returns `false` in all non-affected situations.