README: add information about use with setup-php
As of version 2.15.0 of the `shivammathur/setup-php` action for GitHub Actions, the PHPUnit Polyfills can be installed directly from this action using the `tools` key.

In that case, both this package as well as PHPUnit will be globally installed via Composer.

This commit adds an FAQ section to the readme with a little more information about how this works and how to get things running when installed like that.