Autoload: improve/stabilize version check for TestCase
The version check in the `Autoload::loadTestCase()` method could get confused if the package using this library, or a dependency of the package using this library would alias the `PHPUnit_Runner_Version`/`PHPUnit\Runner\Version` class.

As both classes contain the method we need to retrieve the version number, let's always base the loading on the available PHPUnit version number instead of the availability of either class.