README: update links to PHPUnit docs
As PHPUnit 10.0 has been released, the `main` branch of the PHPUnit docs now points to the PHPUnit 10.0 documentation, while the last release of the Polyfills supports PHPUnit 4.x - 9.x.

This commit updates the links in the README to point explicitly to the PHPUnit 9.6 documentation.

PHPUnit 10.0 support will be handled in a new major of the Polyfills and the links should be updated (again) to point to the PHPUnit 10.0 documentation in the branch containing that major.