PHPUnit 10 | Make all assertion/exception methods `final`
PHPUnit 10.0 makes all assertion and exception methods `final`, so the PHPUnit Polyfills should follow suit.

There is one trait which I've exempted from this: `AssertClosedResource`, as when the methods in that trait (and the empty version of that trait) are made `final`, we'll run into trouble on PHPUnit < 9.x.

Note to self: [PHPCSExtra](https://github.com/PHPCSStandards/PHPCSExtra) 1.1.0 is expected to include a sniff which can enforce this. Once that release is available, that sniff should be added to the PHPCS ruleset to safeguard this.