GH Actions: run tests against PHPUnit PHAR files as well
PHPUnit 8.5.38 and 9.6.19 contain a change in the PHAR build process, which means that classes external to PHPUnit are now prefixed differently, which breaks one polyfill.

As things were, such issues would previously not be caught by CI.

This commit adds a new job to the test workflow, which will run the tests against a range of PHP-PHPUnit combinations using the PHPUnit Phar to ensure any such issues will be automatically caught in the future.

The new test workflow will (very) selectively also upload code coverage to safeguard that the PHAR-file specific lines in the code base remain covered by tests as well.