Tests: improve PHP 8.4 deprecation work-around
Follow-up on PR 159.

This commit builds on the earlier change, but instead of making the whole `AssertObjectEqualsTest` class have a PHP 7.1 minimum requirement, it now limits that requirement to just the one affected test.

The affected test now uses a separate test fixture to allow the bulk of the tests to run on PHP 7.0 (which was the previous minimum requirement for these tests).